src/deep_car/deep_car/requirements.txt
tensorflow-gpu==1.1.0
numpy==1.12.1
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.2.1
h5py==2.7.0
imageio==2.1.1
iso8601==0.1.11
more-itertools==3.2.0
moviepy==0.2.2.11
matplotlib==1.5.1
pandas==0.19.2
Pillow==4.0.0
scikit-image==0.12.3
scipy==0.19.0
seaborn==0.7.1
tqdm==4.11.2

ros-numpy==0.0.2
rosbag==1.11.21
rosboost-cfg==1.11.14
rosclean==1.11.14
roscreate==1.11.14
rosdep==0.11.5
rosdistro==0.6.2
rosdistro-modules==0.6.2
rosgraph==1.11.21
rosinstall==0.7.8
roslaunch==1.11.21
roslib==1.11.14
roslint==0.10.0
roslz4==1.11.21
rosmake==1.11.14
rosmaster==1.11.21
rosmsg==1.11.21
rosnode==1.11.21
rosparam==1.11.21
rospkg==1.1.1
rospkg-modules==1.1.1
rospy==1.11.21
rosservice==1.11.21
rostest==1.11.21
rostopic==1.11.21
rosunit==1.11.14
roswtf==1.11.21

